requirements.txt
streamlit>=1.44.0
requests>=2.32.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
numpy>=2.2.0
pandas>=2.2.0
scikit-learn>=1.6.0
openai>=2.24.0
langchain-openai>=1.1.0
faiss-cpu>=1.10.0
pypdf>=5.0.0
python-docx>=1.1.2
litellm>=1.63.0
ragas>=0.2.10
datasets>=3.0.0
